Fix integer signed vs. unsigned comparison.

+++ nikon/coolpix_a900/ls_sec.h
@@ -30,7 +30,7 @@ private:
     STAGE_5,
   };
   Stage stage_;
-  int index_;
+  uint8_t index_;
 };
 
 enum class Status {
